requirements.txt
annotated-types==0.7.0
asttokens==2.4.1
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
cymem==2.0.8
debugpy==1.8.7
decorator==5.1.1
emoji==2.14.0
en-coreference-web-trf @ https://github.com/explosion/spacy-experimental/releases/download/v0.6.1/en_coreference_web_trf-3.4.0a2-py3-none-any.whl#sha256=9507f2894d60ac3ca89f2ac925dc2222b63f3e8e5998400581b954cfef859654
exceptiongroup==1.2.2
executing==2.1.0
fasttext-wheel==0.9.2
filelock==3.16.1
fsspec==2024.10.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
httplib2==0.22.0
huggingface-hub==0.26.1
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
langcodes==3.4.1
language_data==1.2.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==1.24.3
packaging==24.1
pandas==2.2.3
parso==0.8.4
pathlib_abc==0.1.1
pathy==0.11.0
pillow==10.2.0
platformdirs==4.3.6
praw==7.8.0
prawcore==2.4.0
preshed==3.0.9
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.1.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pydantic==1.8.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.9.11
requests==2.32.3
rich==13.9.2
rsa==4.9
safetensors==0.4.5
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
spacy==3.4.0
spacy-alignments==0.9.1
spacy-experimental==0.6.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.1.9
srsly==2.4.8
stack-data==0.6.3
sympy==1.13.1
thinc==8.1.12
tokenizers==0.13.3
torch==2.5.0+cu124
torchaudio==2.5.0+cu124
torchvision==0.20.0+cu124
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.25.0
typer==0.4.2
typing_extensions==4.12.2
tzdata==2024.2
update-checker==0.18.0
uritemplate==4.1.1
urllib3==2.2.3
wasabi==0.10.1
wcwidth==0.2.13
weasel==0.3.4
websocket-client==1.8.0
wrapt==1.16.0
